Add "get logs" to hook (#24)
* Add "get logs" to hook

Not (yet?) an operator because it is not clear how to _use_ a list-of-dicts
object from XCOM.  So an operator would be very hard to use correctly.

* Specify connection-types in lakeFS provider

Allows DAGs to access the lakeFS hook.

* Prevent overlapping DAG runs of the example lakeFS DAG

This DAG accesses shared state (creates and uses a constant branch name
`example-branch`, then merges it into `main`).  So overlapping runs will
fail.

This should not affect CI, which aways uses a new repo on a new lakeFS
instance.

* [CR] Verify commit IDs _and_ messages in test

Also improve test debuggability: Retry merge sensor sooner to make runs faster.