Add DAG run metadata to commits
Tested by running the sample DAG and seeing desired metadata on web.

Everything in design except note is in there: UI link goes to the DAG run on
Airflow, identifying URL actually GETs the DAG run.

For note we get:
```
[2023-04-24, 14:19:27 UTC] {commit_operator.py:106} WARNING - metadata note not added: $Parent instance <DagRun at 0x7faebfedb2d0> is not bound to a Session; lazy load operation of attribute 'dag_run_note' cannot proceed (Background on this error at: https://sqlalche.me/e/14/bhk3) (possibly undefined fields)
```

Will open an issue to add that; there might be an issue with Airflow not
having a note at that point, or even a bug.